Refactor ForTestingController
