+++ src/main/java/edu/hawaii/its/groupings/exceptions/ExceptionForTesting.java
@@ -1,0 +1,8 @@
+package edu.hawaii.its.groupings.exceptions;
+
+public class ExceptionForTesting extends RuntimeException {
+
+    public ExceptionForTesting(String message) {
+        super(message);
+    }
+}
